main/requirements.txt
pandas
Pillow
numpy
scikit-image
matplotlib
opencv-python
